获取所有蛇信息的方法移到了管理类。
